fix(client): move custom validation to a seperate file
